+++ src/misc/cocitation.c
@@ -24,7 +24,6 @@
 
 #include "igraph_cocitation.h"
 
-#include "igraph_memory.h"
 #include "igraph_adjlist.h"
 #include "igraph_interface.h"